feat(go-pkg): add dynamic route for package metadata and redirects
Added a new route to serve go-import meta tags and redirect users to the appropriate GitHub release or repository. Supports both Go tool requests and browser-based navigation for package discovery.
